Add files via upload
Web-based form in which users enter their names, student IDs, and e-mails, and submit the captions of assigned images. This application is supported by PHP and MySQL for fetching images from the database and storing user data, image names, and captions. It checks for duplication by e-mail. The form will show the images to the user and will provide an input field for the caption.